+++ src/Maestro/Maestro.Contracts/PullRequestActorId.cs
@@ -1,21 +1,39 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
 using System;
 using System.Text;
 using Microsoft.ServiceFabric.Actors;
 
 namespace Maestro.Contracts
 {
+    /// <summary>
+    ///     A factory that creates <see cref="ActorId" /> instances for PullRequestActors
+    /// </summary>
     public static class PullRequestActorId
     {
+        /// <summary>
+        ///     Creates an <see cref="ActorId"/> identifying the PullRequestActor responsible for pull requests for the non-batched subscription
+        ///     referenced by <see paramref="subscriptionId" />.
+        /// </summary>
         public static ActorId Create(Guid subscriptionId)
         {
             return new ActorId(subscriptionId);
         }
 
+        /// <summary>
+        ///     Creates an <see cref="ActorId"/> identifying the PullRequestActor responsible for pull requests for all batched subscriptions
+        ///     targeting the (<see paramref="repository" />, <see paramref="branch" />) pair.
+        /// </summary>
         public static ActorId Create(string repository, string branch)
         {
             return new ActorId(Encode(repository) + ":" + Encode(branch));
         }
 
+        /// <summary>
+        ///   Parses an <see cref="ActorId"/> created by <see cref="Create(string, string)"/> into the (repository, branch) pair that created it.
+        /// </summary>
         public static (string repository, string branch) Parse(ActorId id)
         {
             string str = id.GetStringId();
@@ -23,6 +41,7 @@ namespace Maestro.Contracts
             {
                 throw new ArgumentException("Actor id must be a string kind", nameof(id));
             }
+
             int colonIndex = str.IndexOf(":", StringComparison.Ordinal);
 
             if (colonIndex == -1)
