+++ src/Maestro/Maestro.Web/Program.cs
@@ -57,19 +57,4 @@ namespace Maestro.Web
             ServiceHost.Run(host => host.RegisterStatelessWebService<Startup>("Maestro.WebType"));
         }
     }
-
-    public static class StringExtensions
-    {
-        public static (string left, string right) Split2(this string value, char splitOn)
-        {
-            var idx = value.IndexOf(splitOn);
-
-            if (idx < 0)
-            {
-                return (value, value.Substring(0, 0));
-            }
-
-            return (value.Substring(0, idx), value.Substring(idx + 1));
-        }
-    }
 }
